+++ src/test/java/com/github/teamreflog/reflogserver/acceptance/AcceptanceTest.java
@@ -14,7 +14,7 @@ public abstract class AcceptanceTest {
     @LocalServerPort int port;
 
     @BeforeEach
-    void setUp() {
+    protected void setUp() {
         RestAssured.port = port;
     }
